Find enclosed messages by looking at header
Earlier, we used to read the entire text of a message to determine what
other messages were included within. This is inefficient & may miss
messages even though they were enclosed (due to special characters that
may cause the plain text content to differ).

This change will look at a message's header, which contains a
"References" attribute (in plain text, however), to determine what
messages are enclosed. The header also contains the current message's ID
and the "References" attribute contains the IDs of all the enclosed
messages that can safely be deleted if this message is retained.